requirements.txt
numpy==1.24.4
scikit-image==0.21.0
easyocr==1.7.1
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1+cu118
torchvision
pandas
opencv-python-headless
Pillow
tqdm
pyyaml
natsort
albumentations
nltk